requirements.txt
numpy==1.16.4
pandas==0.24.2
TA-Lib==0.4.17
lightgbm>=2.3.1,<2.4.0
scikit-learn>=0.22.0,<0.23.0
SQLAlchemy>=1.3.6,<1.4.0
PyYAML==5.1
tensorflow>=2.0.0,<2.1.0
